+++ Entity/TB_PositionProvice.cs
@@ -1,7 +1,7 @@
 ﻿//------------------------------------------------------------------------------
 // 
 // 制作人：ChenSheng  
-// 制作日期：2017/06/18
+// 制作日期：2017/06/23
 // 文件说明：省份数据库实体类
 // 
 // 
@@ -23,7 +23,7 @@ namespace CmsEntity
         /// <summary>
         /// 省份id、省份编号
         /// </summary>
-        public int ProviceID { get; set; }
+        public decimal ProviceID { get; set; }
         /// <summary>
         /// 省份名称
         /// </summary>
